Estimate dry density from water and organic content

Estimate dry density from water and organic content

pb210_bulk_density(slice_mass, thickness, core_area = pb210_core_area())

pb210_bulk_density_estimate(
  water_content,
  organic_content,
  density_water = 1000,
  density_organic = 1570,
  density_inorganic = 2830
)

pb210_slice_mass(bulk_density, thickness, core_area = pb210_core_area())

pb210_porosity(
  water_content,
  organic_content,
  density_water = 1000,
  density_organic = 1600,
  density_inorganic = 2830
)

Arguments

slice_mass
The mass of a sediment slice, in kg.

thickness
The thickness of the sediment slice, in m.

core_area
The cross-sectional area being considered. The default is 1 m^2^ such that by default, functions report density and mass on a specific area basis.

water_content
The proportion of the slice made up of water as a proportion of wet mass. Usually this is obtained as mass lost after drying a sample.

organic_content
The proportion of the slice made up of organic material as a proportion of dry mass. Usually this is obtained as mass lost after ignition at 500-600 degrees C (Meyers and Teranes 2001).

density_water
The density of water, in kg/m^3^. Use to adjust this density in the unlikely event that you are dealing with non-standard temperature and/or pressures.

density_organic
The assumed density of (perfectly compacted) organic material. The default is 1570 kg/m^3^, which is a measured value of gyttja particle density as reported by Malloy and Price (2017).

density_inorganic
The assumed density of (perfectly compacted) inorganic material. The default is 2830 kg/m^3^, which is the average crustal density as calculated by Christensen and Mooney (1995).

bulk_density
The density of a sediment slice, in kg/m^3.

References

Christensen, N.I., and Mooney, W.D. 1995. Seismic velocity structure and composition of the continental crust: A global view. Journal of Geophysical Research: Solid Earth, 100: 9761–9788. https://doi.org/10.1029/95JB00259

Malloy, S., and Price, J.S. 2017. Consolidation of gyttja in a rewetted fen peatland: Potential implications for restoration. Mires and Peat,: 1–15. https://doi.org/10.19189/MaP.2015.OMB.200

Meyers, P.A., and Teranes, J.L. 2001. Sediment organic matter. In Tracking Environmental Change Using Lake Sediments: Volume 2: Physical and Geochemical Methods. Edited by W.M. Last and J.P. Smol. Kluwer Academic Publishers, The Netherlands. pp. 239–269. https://doi.org/10.1007/0-306-47670-3_9

Examples

# density of a 10 g slice that is 1 cm thick for a 6.3-cm diameter core
pb210_bulk_density(slice_mass = 0.010, thickness = 0.01, core_area = pb210_core_area(0.063))

#> [1] 320.7961