Apply the Constant Rate of Supply or Constant Initial Concentration model

The CRS model was first published by Robbins (1978) and Appleby and Oldfield (1978), and is behind nearly every lead-210-based age-depth model. These functions compute age-depth models based on excess lead-210 activities (perhaps calculated by pb210_excess()) and propagate error in quadrature using the errors::errors() package. For a more robust estimation of error, consider using pb210_crs_monte_carlo() or pb210_cic_monte_carlo().

pb210_cic(
  cumulative_dry_mass,
  excess,
  model_top = ~pb210_fit_exponential(..1, ..2),
  decay_constant = pb210_decay_constant()
)

pb210_crs(
  cumulative_dry_mass,
  excess,
  inventory = pb210_inventory_calculator(),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant()
)

# S3 method for pb210_fit_cic
predict(object, cumulative_dry_mass = NULL, ...)

# S3 method for pb210_fit_crs
predict(object, cumulative_dry_mass = NULL, ...)

Arguments

cumulative_dry_mass
The cumulative dry mass of the core (in kg), starting at the surface sample and including all samples in the core. These must be greater than 0 and in increasing order.

excess
An excess (non-erosional) lead-210 specific activity (in Bq/kg) for samples where this was measured, and NA where lead-210 was not measured. Use errors::set_errors() to use quadrature error propogation.

model_top
A fit object, such as one generated by pb210_fit_exponential() or a constant specifying the surface excess. The choice of this value has considerable impact on young dates.

decay_constant
The decay contstant for lead-210 (in 1/years). This is an argument rather than a constant because we have found that different spreadsheets in the wild use different decay constants. See pb210_decay_constant().

inventory
The cumulative excess lead-210 activity (in Bq), starting at the bottom of the core. By default, this is estimated by the default pb210_inventory_calculator(). If specifying a vector of values, ensure that the surface (0 cumulative mass) value is specified.

core_area
The internal area of the corer (in m^2^). This can be calculated from an internal diameter using pb210_core_area().

object
A fit object generated by pb210_crs() or pb210_cic().

...
Unused.

Value

predict() methods return a tibble with (at least) components age and age_sd (both in years). CRS model predict() function output also contains inventory, inventory_sd, mar and mar_sd (in kg / m^2^ / year).

References

Appleby, P.G., and Oldfield, F. 1983. The assessment of ^210^Pb data from sites with varying sediment accumulation rates. Hydrobiologia, 103: 29–35. https://doi.org/10.1007/BF00028424

Appleby, P.G., and Oldfield, F. 1978. The calculation of lead-210 dates assuming a constant rate of supply of unsupported ^210^Pb to the sediment. CATENA, 5: 1–8. https://doi.org/10.1016/S0341-8162(78)80002-2

Robbins, J.A. 1978. Geochemical and geophysical applications of radioactive lead isotopes. In The Biogeochemistry of lead in the environment. Edited by J.O. Nriagu. Elsevier/North-Holland Biomedical Press, Amsterdam. pp. 285–393. https://books.google.com/books?id=N4wMAQAAMAAJ

Examples

# simulate a core
core <- pb210_simulate_core() %>%
  pb210_simulate_counting()

# calculate ages using the CRS model
crs <- pb210_crs(
  pb210_cumulative_mass(core$slice_mass),
  set_errors(
    core$activity_estimate,
    core$activity_se
  )
)

#> Error in check_mass_and_activity(cumulative_dry_mass, without_errors(excess)): sum(is.finite(excess) & (excess > 0)) >= 3 is not TRUE

predict(crs)

#> Error in predict(crs): object 'crs' not found