Apply the Constant Rate of Supply or Constant Initial Concentration model

Apply the Constant Rate of Supply or Constant Initial Concentration model

pb210_cic(
  cumulative_dry_mass,
  excess,
  model_top = ~pb210_fit_exponential(..1, ..2),
  decay_constant = pb210_decay_constant()
)

pb210_crs(
  cumulative_dry_mass,
  excess,
  inventory = pb210_inventory_calculator(),
  core_area = pb210_core_area(),
  decay_constant = pb210_decay_constant()
)

# S3 method for pb210_fit_cic
predict(object, cumulative_dry_mass = NULL, ...)

# S3 method for pb210_fit_crs
predict(object, cumulative_dry_mass = NULL, ...)

Arguments

cumulative_dry_mass
The cumulative dry mass of the core (in kg), starting at the surface sample and including all samples in the core. These must be greater than 0 and in increasing order.

excess
An excess (non-erosional) lead-210 specific activity (in Bq/kg) for samples where this was measured, and NA where lead-210 was not measured. Use errors::set_errors() to use automatic error propogation.

model_top
A fit object, such as one generated by pb210_fit_exponential() or a constant specifying the surface excess. The choice of this value has considerable impact on young dates.

decay_constant
The decay contstant for lead-210 (in 1/years). This is an argument rather than a constant because we have found that different spreadsheets in the wild use different decay constants. See pb210_decay_constant().

inventory
The cumulative excess lead-210 activity (in Bq), starting at the bottom of the core. By default, this is estimated by the default pb210_inventory_calculator(). If specifying a vector of values, ensure that the surface (0 cumulative mass) value is specified.

core_area
The internal area of the corer (in m^2^). This can be calculated from an internal diameter using pb210_core_area().

object
A fit object generated by pb210_crs() or pb210_cic().

...
Unused.

Value

predict() methods return a tibble with (at least) components age and age_sd (both in years). CRS model predict() function output also contains inventory, inventory_sd, mar and mar_sd (in kg / m^2^ / year).

References

Appleby, P.G., and Oldfield, F. 1983. The assessment of ^210^Pb data from sites with varying sediment accumulation rates. Hydrobiologia, 103: 29–35. doi:10.1007/BF00028424.

Appleby, P.G., and Oldfield, F. 1978. The calculation of lead-210 dates assuming a constant rate of supply of unsupported ^210^Pb to the sediment. CATENA, 5: 1–8. doi:10.1016/S0341-8162(78)80002-2.

Examples

# simulate a core
core <- pb210_simulate_core() %>%
  pb210_simulate_counting()

# calculate ages using the CIC model
cic <- pb210_cic(
  pb210_cumulative_mass(core$slice_mass),
  set_errors(
    core$activity_estimate,
    core$activity_se
  )
)

predict(cic)

#> # A tibble: 60 x 2
#>      age age_sd
#>    <dbl>  <dbl>
#>  1  2.44  0.197
#>  2  7.82  0.227
#>  3 12.1   0.257
#>  4 17.8   0.305
#>  5 23.1   0.355
#>  6 27.8   0.401
#>  7 33.1   0.456
#>  8 38.7   0.516
#>  9 43.8   0.573
#> 10 49.1   0.634
#> # … with 50 more rows

# calculate ages using the CRS model
crs <- pb210_crs(
  pb210_cumulative_mass(core$slice_mass),
  set_errors(
    core$activity_estimate,
    core$activity_se
  )
)

predict(crs)

#> # A tibble: 60 x 6
#>      age age_sd   mar  mar_sd inventory inventory_sd
#>    <dbl>  <dbl> <dbl>   <dbl>     <dbl>        <dbl>
#>  1  2.40  0.698 0.151 0.00373     2989.         65.1
#>  2  7.40  0.712 0.153 0.00384     2557.         56.5
#>  3 12.5   0.728 0.149 0.00380     2183.         48.9
#>  4 17.6   0.752 0.152 0.00395     1859.         42.4
#>  5 22.7   0.779 0.153 0.00408     1585.         36.9
#>  6 27.9   0.810 0.151 0.00411     1349.         32.1
#>  7 33.1   0.847 0.151 0.00424     1145.         27.9
#>  8 38.4   0.891 0.152 0.00443      972.         24.4
#>  9 43.7   0.937 0.151 0.00456      824.         21.4
#> 10 49.0   0.988 0.151 0.00474      697.         18.7
#> # … with 50 more rows